Moved LineChart component to a dedicated visualizations directory.
